Took away messages being sent to inverters when RTD switch is off to fix inverter startup.  Uncommented TS latching code in precharge.c. Added software debounce in pio.c. In the handle_pio function, the state ts_active only updates after the TS_INPUT_Get() has been different from ts_active for 30 ms.

+++ src/inverter.h
@@ -21,7 +21,7 @@ extern inv_t inv1;
 extern inv_t inv2;
 
 int inv_parse_rx(volatile char* msg, volatile size_t len, inv_t* inv, size_t (*io_write)(uint8_t*,const size_t) ); 
-uint16_t get_inv_lowest_voltage(void);
+float get_inv_lowest_voltage(void);
 
 void handle_inverter(void);
